site/src/main/webapp/app/entities/machine-documentation/route/machine-documentation-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IMachineDocumentation } from '../machine-documentation.model';
import { MachineDocumentationService } from '../service/machine-documentation.service';

import machineDocumentationResolve from './machine-documentation-routing-resolve.service';

describe('MachineDocumentation routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let service: MachineDocumentationService;
  let resultMachineDocumentation: IMachineDocumentation | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    service = TestBed.inject(MachineDocumentationService);
    resultMachineDocumentation = undefined;
  });

  describe('resolve', () => {
    it('should return IMachineDocumentation returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        machineDocumentationResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultMachineDocumentation = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMachineDocumentation).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      TestBed.runInInjectionContext(() => {
        machineDocumentationResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultMachineDocumentation = result;
          },
        });
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultMachineDocumentation).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IMachineDocumentation>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        machineDocumentationResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultMachineDocumentation = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMachineDocumentation).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
